fix(os): decouple packaged relay sanitizer
