+++ plugins/plugin-agent-orchestrator/__tests__/shared-runtime-env.ts
@@ -4,3 +4,12 @@ export {
   resolvePlatform,
 } from "../../../packages/shared/src/runtime-env.js";
 export { readAliasedEnv } from "../../../packages/shared/src/utils/env.js";
+export {
+  DEFAULT_MAX_RELAY_CHARS,
+  elideLongBlocks,
+  sanitizeCompletionRelay,
+  stripEnvelopeSummaryLines,
+  stripStructuredProofLines,
+  stripToolTranscript,
+  TOOL_OUTPUT_END_MARKER,
+} from "../../../packages/shared/src/utils/transcript-sanitizer.js";
